Revert "moved add_framework_metadata"
This reverts commit 4e9703d1f95bd8556bf7b44b75b56034b93af1da.
